Change modules structure; tests refactoring
