+++ signit/__init__.py
@@ -1,6 +1,4 @@
-from .signature import generate_key
-from .signature import create_signature
-from .signature import parse_signature
-from .signature import verify_signature
+import key
+import signature
 
 __version__ = '0.1.3'
